Refactored app logic, helper functions now in chaliclib
